Move variant_qc info root large_cohort module
